refactor lots
